+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Increased damage taken by wither from blue skulls
   * Damage from blue skulls is multiplied by 0.8% -> 1% of its max health
+* Renamed default wither stats names in data pack
 * Charge Object
   * Added `second_phase` in Wither's charge object. If true, the wither will charge 3 times in rapid succession
   * Added `second_phase_times`, `second_phase_tick_reduction` and `second_phase_max_reduction` to configure the second phase charging times, tick reduction and max reduction